APIs/beezup.com/2.0/python-aiohttp-client/tests/test_customer_shares_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.beez_up_common_error_response_message import BeezUPCommonErrorResponseMessage
from openapi_server.models.error_response_message_payment_required import ErrorResponseMessagePaymentRequired
from openapi_server.models.store_shares import StoreShares


pytestmark = pytest.mark.asyncio

async def test_delete_store_share(client):
    """Test case for delete_store_share

    Delete a share of a store to another user
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='DELETE',
        path='/v2/user/customer/stores/{store_id}/shares/{user_id}'.format(store_id='store_id_example', user_id='user_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_store_shares(client):
    """Test case for get_store_shares

    Get shares related to this store
    """
    headers = { 
        'Accept': 'application/json',
        'if_none_match': 'if_none_match_example',
    }
    response = await client.request(
        method='GET',
        path='/v2/user/customer/stores/{store_id}/shares'.format(store_id='store_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_share_store(client):
    """Test case for share_store

    Share a store to another user
    """
    body = 'body_example'
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
    }
    response = await client.request(
        method='POST',
        path='/v2/user/customer/stores/{store_id}/shares'.format(store_id='store_id_example'),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

